+++ src/quiz/quiz_command_handler.js
@@ -4,7 +4,7 @@ const QuizJSON      = require("../../" + questionsFile);
 const CommandHandlerBase = require("../command_handler_base")
 const Config        = require('../../data/config.json');
 const Bot           = require("../main");
-const Util          = require("../misc/util");
+const Util          = require("../util");
 const QuizSession   = require("./quiz_session")
 
 const fs            = require('fs');
